feat: identidade PBH, workflow zeladoria, parametrização e preparação pre-launch
Tema institucional PBH (azul marinho #1B2A6B) aplicado a todo o sistema.
Cards em vez de tabelas nas listagens de vistorias e pontos.
Workflow semântico de zeladoria com headers por fase (Atualizar Informação,
Retorno Agendado, Retorno Vencido, Concluída) e cores distintas por ação.

Conceito de Informação Precária: pontos sem vistoria há +60 dias (parametrizável).
Pesos de complexidade individuais por fator (16 fatores, VI-SPDAT).
29 parâmetros configuráveis via /admin/parametros com abas por grupo.

Mapa restrito a Belo Horizonte (maxBounds + viscosity 1.0).
Mobilidade de pessoas entre pontos (busca + vinculação na vistoria).
Nomenclatura corrigida: "Pessoas em Situação de Rua" em toda a interface.

CRUD de vistoria equalizado entre create e edit (moradores, toggles, stepper).
Policy: owner e admin podem editar qualquer vistoria (homologação).
Seeder de permissões para produção (23 permissões, 7 roles).
Throttle em login e API. Inline handlers migrados para Alpine.js.

Documentos: nota técnica, proposta de modelo de dados, levantamento de requisitos.
Portaria Conjunta 009/2026 capturada e referenciada.

376 testes, 951 asserções, PHPStan L6 zero erros, Pint limpo.

+++ app/Services/DashboardService.php
@@ -19,7 +19,7 @@ class DashboardService
     /** Total de pontos cadastrados */
     public function totalPontos(): int
     {
-        return Cache::remember(
+        return (int) Cache::remember(
             'dashboard:total_pontos',
             self::CACHE_TTL,
             fn () => Ponto::query()->count()
